Changed Problem 19 to use floating point variables instead of integers as shown in the course

+++ Problem Set 1/prob19.c
@@ -2,16 +2,16 @@
 
 int main()
 {
-    int a;
-    int b;
-    int sum;
-    int mul;
-    int sub;
-    int div;
+    float a;
+    float b;
+    float sum;
+    float mul;
+    float sub;
+    float div;
     int choice;
 
     printf("Enter 2 numbers(a b): ");
-    scanf("%d %d", &a, &b);
+    scanf("%f %f", &a, &b);
 
     printf("Choose 1 of the few: \n\n1)Addition\n2)Multiplication\n3)Subtraction\n4)Division\n\n");
     scanf("%d", &choice);
@@ -20,19 +20,19 @@ int main()
     {
         case 1:
             sum=a+b;
-            printf("The sum of %d and %d is %d\n", a, b, sum);
+            printf("The sum of %.2f and %.2f is %.2f\n", a, b, sum);
             break;
         case 2:
             mul=a*b;
-            printf("The product of %d and %d is %d\n", a, b, mul);
+            printf("The product of %.2f and %.2f is %.2f\n", a, b, mul);
             break;
         case 3:
             sub=a-b;
-            printf("The difference of %d and %d is %d\n", a, b, sub);
+            printf("The difference of %.2f and %.2f is %.2f\n", a, b, sub);
             break;
         case 4:
             div=a/b;
-            printf("The quotient of %d and %d is %d\n", a, b, div);
+            printf("The quotient of %.2f and %.2f is %.2f\n", a, b, div);
             break;
     }
 }
